Move ElasticSearchBoostConfig into settings
ECOM-7079
